+++ src/ComboboxInput/ComboboxInput.js
@@ -1,26 +1,46 @@
+import classnames from 'classnames';
 import { Popover } from '../Popover/Popover';
 import PropTypes from 'prop-types';
 import React from 'react';
 
 // ------------------------------------------- Combobox Input ------------------------------------------
-export const ComboboxInput = ({ placeholder, menu, compact, className, ...props }) => {
+export const ComboboxInput = ({ placeholder, menu, compact, className, popoverProps, inputProps, buttonProps, ...props }) => {
+    const comboboxInputClasses = classnames(
+        'fd-combobox-input',
+        className
+    );
+
+    const comboboxPopoverClasses = classnames(
+        'fd-input-group',
+        'fd-input-group--after',
+        {
+            'fd-input-group--compact': compact
+        }
+    );
+
+    const comboboxPopoverInputClasses = classnames(
+        'fd-input',
+        {
+            'fd-input--compact': compact
+        }
+    );
+
     return (
-        <div className={`fd-combobox-input${className ? ' ' + className : ''}`} {...props}>
+        <div {...props} className={comboboxInputClasses}>
             <Popover
+                {...popoverProps}
                 body={menu}
                 control={
                     <div className='fd-combobox-control'>
                         <div
-                            className={`fd-input-group fd-input-group--after${
-                                compact ? ' fd-input-group--compact' : ''
-                            }`}>
+                            className={comboboxPopoverClasses}>
                             <input
-                                className={`fd-input${compact ? ' fd-input--compact' : ''}`}
-                                id=''
+                                {...inputProps}
+                                className={comboboxPopoverInputClasses}
                                 placeholder={placeholder}
                                 type='text' />
                             <span className='fd-input-group__addon fd-input-group__addon--after fd-input-group__addon--button'>
-                                <button className=' fd-button--light sap-icon--navigation-down-arrow' />
+                                <button {...buttonProps} className=' fd-button--light sap-icon--navigation-down-arrow' />
                             </span>
                         </div>
                     </div>
@@ -32,9 +52,12 @@ export const ComboboxInput = ({ placeholder, menu, compact, className, ...props 
 
 ComboboxInput.propTypes = {
     menu: PropTypes.object.isRequired,
+    buttonProps: PropTypes.object,
     className: PropTypes.string,
     compact: PropTypes.bool,
-    placeholder: PropTypes.string
+    inputProps: PropTypes.object,
+    placeholder: PropTypes.string,
+    popoverProps: PropTypes.object
 };
 
 ComboboxInput.defaultTypes = {