+++ src/MegaMenu/MegaMenu.js
@@ -1,12 +1,18 @@
+import classnames from 'classnames';
 import PropTypes from 'prop-types';
 import { BrowserRouter, Link } from 'react-router-dom';
 import React, { Component } from 'react';
 
 export const MegaMenu = ({ children, className, ...props }) => {
+    const megaMenuClasses = classnames(
+        'fd-mega-menu',
+        className
+    );
+
     return (
         <nav
-            className={`fd-mega-menu${className ? ' ' + className : ''}`}
-            {...props}>
+            {...props}
+            className={megaMenuClasses}>
             {children}
         </nav>
     );
@@ -38,114 +44,120 @@ export class MegaMenuList extends Component {
         };
     }
 
-  handleSelectChild = (e, id) => {
-      this.setState({
-          selectedItem: id
-      });
-  };
+    handleSelectChild = (e, id) => {
+        this.setState({
+            selectedItem: id
+        });
+    };
 
-  handleSelect = (e, id) => {
-      const { itemStates } = this.state;
-      let iStates = itemStates;
-      iStates[id] = !iStates[id];
-      Object.keys(iStates).map((key) => {
-          if (key === id) {
-              iStates[key] = true;
-          } else {
-              iStates[key] = false;
-          }
-      });
-      this.setState({ itemStates: iStates });
-      this.setState({ selectedItem: id });
-  };
+    handleSelect = (e, id) => {
+        const { itemStates } = this.state;
+        let iStates = itemStates;
+        iStates[id] = !iStates[id];
+        Object.keys(iStates).map((key) => {
+            if (key === id) {
+                iStates[key] = true;
+            } else {
+                iStates[key] = false;
+            }
+        });
+        this.setState({ itemStates: iStates });
+        this.setState({ selectedItem: id });
+    };
 
-  render() {
-      const { items, className, ...props } = this.props;
-      return (
-          <BrowserRouter>
-              <ul {...props} className={`fd-mega-menu__list${className ? ' ' + className : ''}`}>
-                  {items.map(item => {
-                      return (
-                          <li className='fd-mega-menu__item' key={item.id}>
-                              {item.link ? (
-                                  <Link
-                                      className={`fd-mega-menu__link${
-                                          this.state.selectedItem === item.id ? ' is-selected' : ''
-                                      }${item.hasChild ? ' has-child' : ''}${
-                                          this.state.itemStates[item.id] && item.hasChild
-                                              ? ' is-expanded'
-                                              : ''
-                                      }`}
-                                      key={item.id}
-                                      onClick={e => this.handleSelect(e, item.id)}
-                                      to={{ pathname: item.link }}>
-                                      {item.name}
-                                  </Link>
-                              ) : null}
+    getLinkClasses = ({id, hasChild}) => {
+        return classnames(
+            'fd-mega-menu__link',
+            {
+                'is-selected': this.state.selectedItem === id,
+                'has-child': hasChild,
+                'is-expanded': this.state.itemStates[id] && hasChild
+            }
+        );
+    }
 
-                              {item.url ? (
-                                  <a
-                                      className={`fd-mega-menu__link${
-                                          this.state.selectedItem === item.id ? ' is-selected' : ''
-                                      }${item.hasChild ? ' has-child' : ''}${
-                                          this.state.itemStates[item.id] && item.hasChild
-                                              ? ' is-expanded'
-                                              : ''
-                                      }`}
-                                      href={item.url}
-                                      key={item.id}
-                                      onClick={e => this.handleSelect(e, item.id)}>
-                                      {item.name}
-                                  </a>
-                              ) : null}
+    getSubLinkClasses = ({id}) => {
+        return classnames(
+            'fd-mega-menu__sublink',
+            {
+                'is-selected': this.state.selectedItem === id
+            }
+        );
+    }
 
-                              {item.hasChild ? (
-                                  <ul
-                                      aria-expanded={this.state.itemStates[item.id]}
-                                      aria-hidden={!this.state.itemStates[item.id]}
-                                      className='fd-mega-menu__sublist'
-                                      id={item.id}>
-                                      {item.child.map(ch => {
-                                          return (
-                                              <li className='fd-mega-menu__subitem' key={ch.id}>
-                                                  {ch.link ? (
-                                                      <Link
-                                                          className={`fd-mega-menu__sublink${
-                                                              this.state.selectedItem === ch.id
-                                                                  ? ' is-selected'
-                                                                  : ''
-                                                          }`}
-                                                          key={ch.id}
-                                                          onClick={e => this.handleSelectChild(e, ch.id)}
-                                                          to={{ pathname: ch.link }}>
-                                                          {ch.name}
-                                                      </Link>
-                                                  ) : null}
-                                                  {ch.url ? (
-                                                      <a
-                                                          className={`fd-mega-menu__sublink${
-                                                              this.state.selectedItem === ch.id
-                                                                  ? ' is-selected'
-                                                                  : ''
-                                                          }`}
-                                                          href={ch.url}
-                                                          key={ch.id}
-                                                          onClick={e => this.handleSelectChild(e, ch.id)}>
-                                                          {ch.name}
-                                                      </a>
-                                                  ) : null}
-                                              </li>
-                                          );
-                                      })}
-                                  </ul>
-                              ) : null}
-                          </li>
-                      );
-                  })}
-              </ul>
-          </BrowserRouter>
-      );
-  }
+    render() {
+        const { items, className, ...props } = this.props;
+
+        const megaMenuListClasses = classnames(
+            'fd-mega-menu__list',
+            className
+        );
+
+        return (
+            <BrowserRouter>
+                <ul {...props} className={megaMenuListClasses}>
+                    {items.map(item => {
+                        return (
+                            <li className='fd-mega-menu__item' key={item.id}>
+                                {item.link ? (
+                                    <Link
+                                        className={this.getLinkClasses(item)}
+                                        key={item.id}
+                                        onClick={e => this.handleSelect(e, item.id)}
+                                        to={{ pathname: item.link }}>
+                                        {item.name}
+                                    </Link>
+                                ) : null}
+
+                                {item.url ? (
+                                    <a
+                                        className={this.getLinkClasses(item)}
+                                        href={item.url}
+                                        key={item.id}
+                                        onClick={e => this.handleSelect(e, item.id)}>
+                                        {item.name}
+                                    </a>
+                                ) : null}
+
+                                {item.hasChild ? (
+                                    <ul
+                                        aria-expanded={this.state.itemStates[item.id]}
+                                        aria-hidden={!this.state.itemStates[item.id]}
+                                        className='fd-mega-menu__sublist'
+                                        id={item.id}>
+                                        {item.child.map(ch => {
+                                            return (
+                                                <li className='fd-mega-menu__subitem' key={ch.id}>
+                                                    {ch.link ? (
+                                                        <Link
+                                                            className={this.getSubLinkClasses(ch.id)}
+                                                            key={ch.id}
+                                                            onClick={e => this.handleSelectChild(e, ch.id)}
+                                                            to={{ pathname: ch.link }}>
+                                                            {ch.name}
+                                                        </Link>
+                                                    ) : null}
+                                                    {ch.url ? (
+                                                        <a
+                                                            className={this.getSubLinkClasses(ch.id)}
+                                                            href={ch.url}
+                                                            key={ch.id}
+                                                            onClick={e => this.handleSelectChild(e, ch.id)}>
+                                                            {ch.name}
+                                                        </a>
+                                                    ) : null}
+                                                </li>
+                                            );
+                                        })}
+                                    </ul>
+                                ) : null}
+                            </li>
+                        );
+                    })}
+                </ul>
+            </BrowserRouter>
+        );
+    }
 }
 MegaMenuList.propTypes = {
     items: PropTypes.array.isRequired,
@@ -153,8 +165,13 @@ MegaMenuList.propTypes = {
 };
 
 export const MegaMenuGroup = ({ title, children, className, headerProps, ...props }) => {
+    const megaMenuGroupClasses = classnames(
+        'fd-mega-menu__group',
+        className
+    );
+
     return (
-        <div {...props} className={`fd-mega-menu__group${className ? ' ' + className : ''}`}>
+        <div {...props} className={megaMenuGroupClasses}>
             <h1 {...headerProps} className='fd-mega-menu__title'>{title}</h1>
             {children}
         </div>