+++ src/Table/Table.js
@@ -1,21 +1,43 @@
+import classnames from 'classnames';
 import PropTypes from 'prop-types';
 import React from 'react';
 
 export const Table = props => {
-    const { headers, tableData, className, ...rest } = props;
+    const {
+        headers,
+        tableData,
+        className,
+        tableBodyProps,
+        tableBodyRowProps,
+        tableHeaderProps,
+        tableHeaderRowProps,
+        ...rest
+    } = props;
+
+    const tableClasses = classnames(
+        'fd-table',
+        className
+    );
+
     return (
-        <table className={`fd-table${className ? ' ' + className : ''}`} {...rest}>
-            <thead>
-                <tr>
+        <table {...rest} className={tableClasses}>
+            <thead {...tableHeaderProps}>
+                <tr {...tableHeaderRowProps}>
                     {headers.map((header, index) => {
                         return <th key={index}>{header}</th>;
                     })}
                 </tr>
             </thead>
-            <tbody>
+            <tbody {...tableBodyProps}>
                 {tableData.map((row, index) => {
+                    let rowProps;
+                    if (tableBodyRowProps) {
+                        rowProps = (typeof tableBodyRowProps === 'function'
+                            ? tableBodyRowProps(row, index)
+                            : tableBodyRowProps);
+                    }
                     return (
-                        <tr key={index}>
+                        <tr {...rowProps} key={index}>
                             {row.rowData.map((rowData, cellIndex) => {
                                 return <td key={cellIndex}>{rowData}</td>;
                             })}
@@ -33,5 +55,12 @@ Table.propTypes = {
         }).isRequired
     ).isRequired,
     className: PropTypes.string,
-    headers: PropTypes.array
+    headers: PropTypes.array,
+    tableBodyProps: PropTypes.object,
+    tableBodyRowProps: PropTypes.oneOfType([
+        PropTypes.object,
+        PropTypes.func
+    ]),
+    tableHeaderProps: PropTypes.object,
+    tableHeaderRowProps: PropTypes.object
 };