+++ src/Tree/Tree.Component.js
@@ -106,7 +106,10 @@ export const TreeComponent = () => {
 
             <Properties properties={[
                 { name: 'headers', description: 'array of strings for the column headers of the tree' },
-                { name: 'treeData', description: 'array of objects that contain several properties, id (the id of the row), values (an array of strings containing data for each column in the row), hasChildren(a boolean value whether the row contains children or not) and children (an array of objects containing additional rows).' }
+                { name: 'treeData', description: 'array of objects that contain several properties, id (the id of the row), values (an array of strings containing data for each column in the row), hasChildren(a boolean value whether the row contains children or not) and children (an array of objects containing additional rows).' },
+                { name: 'headerProps', description: 'object - additional props to be spread to the header element.' },
+                { name: 'headerButtonProps', description: 'object - additional props to be spread to the header expand button element.' },
+                { name: 'listProps', description: 'object - additional props to be spread to the list element.' }
             ]} type='Inputs' />
 
             <Separator />
@@ -188,12 +191,14 @@ export const TreeComponent = () => {
             <Properties properties={[
                 { name: 'displayText', description: 'The text to display in the cell. If omitted, the link url will be displayed.' },
                 { name: 'linkUrl', description: 'If provided, this is the URL the link will navigate to.' }
-
             ]} type='Properties' />
+
             <Properties properties={[
                 { name: 'headers', description: 'Array of strings for the column headers of the tree' },
-                { name: 'treeData', description: '	Array of objects that contain several properties, id(the id of the row), hasChildren(a boolean value whether the row contains children or not), values (an array of objects containing data for each column in the row), and children (an array of objects containing additional rows).' }
-
+                { name: 'treeData', description: '	Array of objects that contain several properties, id(the id of the row), hasChildren(a boolean value whether the row contains children or not), values (an array of objects containing data for each column in the row), and children (an array of objects containing additional rows).' },
+                { name: 'headerProps', description: 'object - additional props to be spread to the header element.' },
+                { name: 'headerButtonProps', description: 'object - additional props to be spread to the header expand button element.' },
+                { name: 'listProps', description: 'object - additional props to be spread to the list element.' }
             ]} type='Inputs' />
             <Separator />
             <DocsTile>